App/requirements.txt
flask
peewee
faiss-cpu
numpy
gradio
langchain-community
sentence-transformers
python-dotenv
streamlit
together
